Create 06aa74d2701b_.py
So now we have a migration file that contains Python code that knows how to create our database structure from scratch. It's worth noting that the file only contains information about the database structure -- that is, there's a table called comments that has certain specific columns. It doesn't contain the comments themselves. Take a look at the file -- it was the last thing printed out by the flask db migrate command, and you can load it up into the editor via the "Files" page. It will look something like this: